rename main.py to moscg.py
